optimize Home page components 🤩 " Unfortunately, I was not able to compile the home properly, but you can do it
