+++ pages/Home/components/SectionOneAnnounce.jsx
@@ -3,8 +3,8 @@ import { React } from "react";
 import Image from "next/image";
 
 import Lottie from "lottie-react";
-import celebrate from "../public/JSON/user/60051-shapes-v1.json";
-import celebrate2 from "../public/JSON/user/60061-shapes-gradient";
+import celebrate from "../../../public/JSON/user/60051-shapes-v1.json";
+import celebrate2 from "../../../public/JSON/user/60061-shapes-gradient.json";
 import { useRouter } from "next/router";
 import { Loading } from "@nextui-org/react";
 import Link from "next/link";
@@ -12,7 +12,7 @@ import Link from "next/link";
 function SectionOneAnnounce(props) {
 
   let router = useRouter();
-let distnationPartner = props.partner;
+  let distnationPartner = props.partner;
   return (
     <div
       style={{ fontFamily: "Poppins, sans-serif" }}
@@ -21,46 +21,41 @@ let distnationPartner = props.partner;
       }>
       <div className='relative col-span-3 '>
         <div className='sm:ml-7  text-center  md:text-left max-w-3xl mx-auto pt-32 sm:pt-60 md:pt-96 lg:pt-60'>
-          <h1       style={{ fontFamily: "Poppins-Bold" }}
- className='select-none  text-slate-200 tracking-tight  leading-loose
- 
-
- font-extrabold text-4xl sm:text-5xl lg:text-6xl  text-center lg:text-left '>
+          <h1 style={{ fontFamily: "Poppins-Bold" }}
+              className='select-none  text-slate-200 tracking-tight  leading-loose font-extrabold text-4xl sm:text-5xl lg:text-6xl  text-center lg:text-left '>
             {props.MainText}
             <span className=' text-red-500  font-extrabold  text-4xl sm:text-5xl lg:text-6xl tracking-tight text-center  lg:text-left '>
               .
             </span>
           </h1>
-          <p
-            style={{ fontFamily: "Poppins-Bold", width: "400" }}
-            className='mt-6 font-medium  tracking-tight text-md sm:text-xl  text-center  lg:text-left  max-w-3xl mx-7 text-slate-400'>
-            {props.MainPTexe}
-          </p>
-          <p
-            style={{ fontFamily: "Poppins-Bold", width: "400" }}
-            className='mt-6 font-extralight	  tracking-tight text-sm sm:text-lg  text-center  lg:text-left  max-w-3xl mx-7 text-slate-400/60'>
-            {props.MainPTexe2}
-          </p>
+            <p
+              style={{ fontFamily: "Poppins-Bold", width: "400" }}
+              className='mt-6 font-medium  tracking-tight text-md sm:text-xl  text-center  lg:text-left  max-w-3xl mx-7 text-slate-400'>
+              {props.MainPTexe}
+            </p>
+            <p
+              style={{ fontFamily: "Poppins-Bold", width: "400" }}
+              className='mt-6 font-extralight	  tracking-tight text-sm sm:text-lg  text-center  lg:text-left  max-w-3xl mx-7 text-slate-400/60'>
+              {props.MainPTexe2}
+            </p>
 
           <div className='mt-6  mx-12 sm:mx-0   sm:mt-10  sm:flex justify-center  md:justify-center lg:justify-start  sm:space-x-6 text-sm'>
             <Link legacyBehavior
               aria-label='Get Started'
               href={props.action}
               passHref>
-                <div               className='select-none max-w-sm font-bold box-decoration-slice shadow-lg hover:shadow-red-500/50     focus:outline-none focus:ring-2 focus:ring-slate-400 focus:ring-offset-2 active:scale-95 duration-100 focus:ring-offset-slate-800   h-12 px-6 rounded-lg w-full flex items-center text-white justify-center sm:w-auto bg-red-500  hover:bg-red-400'
->
-{props.ActionButton}
-  
-</div>
+                <div className='select-none max-w-sm font-bold box-decoration-slice shadow-lg hover:shadow-red-500/50     focus:outline-none focus:ring-2 focus:ring-slate-400 focus:ring-offset-2 active:scale-95 duration-100 focus:ring-offset-slate-800   h-12 px-6 rounded-lg w-full flex items-center text-white justify-center sm:w-auto bg-red-500  hover:bg-red-400'>
+                   {props.ActionButton}
+              </div>
            </Link>
             <Link legacyBehavior
               aria-label='Our Services'
               href={props.action2}
               passHref>
-                <div               className='select-none font-bold max-w-sm  bg-white mt-4 sm:mt-0 shadow-md  hover:shadow-red-500/50  ring-2  ring-slate-900 hover:bg-gray-300 focus:outline-none focus:ring-2 opacity-90  focus:ring-offset-2 active:scale-95 duration-100 0  text-gray-900  h-12 px-3  sm:px-16 rounded-lg w-full flex items-center justify-center sm:w-auto highlight-white/20 '
->             {props.ActionButton2}
- </div>
- 
+                <div className='select-none font-bold max-w-sm  bg-white mt-4 sm:mt-0 shadow-md  hover:shadow-red-500/50  ring-2  ring-slate-900 hover:bg-gray-300 focus:outline-none focus:ring-2 opacity-90  focus:ring-offset-2 active:scale-95 duration-100 0  text-gray-900  h-12 px-3  sm:px-16 rounded-lg w-full flex items-center justify-center sm:w-auto highlight-white/20 '>
+                  {props.ActionButton2}
+               </div>
+
             </Link>
           </div>
         </div>
@@ -138,27 +133,24 @@ let distnationPartner = props.partner;
                 </div>
               </div>
             </div>
-            
+
               <div onClick={()=>{
-                
+
                 router.push(distnationPartner)
               }} className='h-56 w-56 relative backdrop-blur-xl bg-white/30 cursor-pointer	 opacity-90 hover:opacity-50 duration-500'>
                 {" "}
                 <div
-                  className='backdrop-blur-xl saturate-50
-
-bg-white/30 p-2 hover:p-12 duration-500'>
+                  className='backdrop-blur-xl saturate-50 bg-white/30 p-2 hover:p-12 duration-500'>
                   <Lottie animationData={celebrate2} />{" "}
-
                   <div className=' top-[40%] text-center flex justify-center absolute '>
                     <h1 className=' ml-2 font-bold text-2xl drop-shadow-2xl bg-gray-900 text-amber-400 select-none  '>
                       Be A Speaker
                     </h1>
-                  </div>   
+                  </div>
                 </div>
-             
+
               </div>
-            
+
           </div>
         </div>
       </div>
